requirements.txt
pytest
allure-pytest==2.13.5
selene>=2.0.0rc9